+++ src/8thWonderland/Application/Controller/IntranetController.php
@@ -18,7 +18,8 @@ class IntranetController extends ActionController {
             'avatar' => $member->getAvatar(),
             'admin' => $this->application->get('member_manager')->isMemberInGroup($member, 1),
             'motions' => $this->application->get('motion_manager')->getActiveMotions($member),
-            'groups' => $member->getGroups()
+            'groups' => $member->getGroups(),
+            'nb_unread_messages' => $this->application->get('message_manager')->countUnreadMessages($member->getId())
         ]);
     }